Moved Projects scenes.
